# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: api
  version: 1.0.0
  description: Shawrome Car Sales Platform API
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    RefreshResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
    ClientRegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string

    # ──────────────────────────── User ────────────────────────────
    UserRole:
      type: string
      enum: [admin, seller, client]
    User:
      type: object
      required: [id, name, email, role, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    CarStatus:
      type: string
      enum: [available, reserved, sold]
    CarFuel:
      type: string
      enum: [essence, diesel, hybride, electrique, gpl]
    CarTransmission:
      type: string
      enum: [manuelle, automatique]
    CarCondition:
      type: string
      enum: [neuf, occasion]
    Car:
      type: object
      required: [id, brand, model, year, mileage, purchasePrice, sellingPrice, fuel, transmission, color, condition, status, photos, sellerId, sellerName, createdAt]
      properties:
        id:
          type: integer
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        description:
          type: string
          nullable: true
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        fuel:
          $ref: '#/components/schemas/CarFuel'
        transmission:
          $ref: '#/components/schemas/CarTransmission'
        color:
          type: string
        condition:
          $ref: '#/components/schemas/CarCondition'
        status:
          $ref: '#/components/schemas/CarStatus'
        photos:
          type: array
          items:
            type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        soldAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    PriceHistory:
      type: object
      required: [id, carId, carBrand, carModel, oldPrice, newPrice, modifiedBy, modifiedByName, createdAt]
      properties:
        id:
          type: integer
        carId:
          type: integer
        carBrand:
          type: string
        carModel:
          type: string
        oldPrice:
          type: number
        newPrice:
          type: number
        modifiedBy:
          type: integer
        modifiedByName:
          type: string
        createdAt:
          type: string
          format: date-time
    PriceHistoryListResponse:
      type: object
      required: [history, total]
      properties:
        history:
          type: array
          items:
            $ref: '#/components/schemas/PriceHistory'
        total:
          type: integer

    # ──────────────────────────── Sales ────────────────────────────
...
    Favorite:
      type: object
      required: [id, carId, car, createdAt]
      properties:
        id:
          type: integer
        carId:
          type: integer
        car:
          $ref: '#/components/schemas/Car'
        createdAt:
          type: string
          format: date-time
    FavoritesListResponse:
      type: object
      required: [favorites]
      properties:
        favorites:
          type: array
          items:
            $ref: '#/components/schemas/Favorite'
    CreateFavoriteRequest:
      type: object
      required: [carId]
      properties:
        carId:
          type: integer
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /client/register:
    post:
      operationId: clientRegister
      tags: [client]
      summary: Register client account (mobile)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRegisterRequest'
      responses:
        '201':
          description: Client registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /client/login:
    post:
      operationId: clientLogin
      tags: [client]
      summary: Client login (mobile)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  # ──────────────────────────── Cars ────────────────────────────
...
  /price-history:
    get:
      operationId: listPriceHistory
      tags: [priceHistory]
      security:
        - bearerAuth: []
      summary: List all price changes (admin)
      parameters:
        - name: carId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Price history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceHistoryListResponse'
  /price-history/{carId}:
    get:
      operationId: getCarPriceHistory
      tags: [priceHistory]
      security:
        - bearerAuth: []
      summary: Get price history for a specific car
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Price history for car
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceHistoryListResponse'

  # ──────────────────────────── Sales ────────────────────────────
...
  /favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      security:
        - bearerAuth: []
      summary: Get current client's favorites
      responses:
        '200':
          description: Favorites list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoritesListResponse'

    post:
      operationId: addFavorite
      tags: [favorites]
      security:
        - bearerAuth: []
      summary: Add car to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFavoriteRequest'
      responses:
        '201':
          description: Added to favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
  /favorites/{carId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      security:
        - bearerAuth: []
      summary: Remove car from favorites
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Removed from favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # ──────────────────────────── Export ────────────────────────────
...