# lib/api-zod/src/generated/types/favoritesListResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Shawrome Car Sales Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { Favorite } from './favorite';

export interface FavoritesListResponse {
  favorites: Favorite[];
}